removed description other stuff and shit (#35)
* added transaction history (#22)

* added dashboard

* added AppDrawer component

* added transaction history

---------


* Database functions (#24)

* database functions

* initial function testing

* partial exception functions

---------


* image test working init on product upload

* still testing product image thingy

* save image post final

* post backend integration for products

* dropdown thing

* remove description prdocut

---------

+++ lib/models/ProductsModel.dart
@@ -2,7 +2,7 @@ class Product {
   final String id;
   final String name;
   final double selling_price;
-  final int total_purchase;
+  final double total_purchase;
   final int product_stock;
   final String category;
   final String image;
